requirements.txt
langchain-community==0.3.*
langchain-openai==0.1.*
langchain-text-splitters==0.1.*
chromadb>=0.5
python-dotenv>=1.0
openai>=1.12          # cliente oficial
